HSEARCH-3399 Introduce a AbstractStandardLuceneFieldSortBuilderFactory
