+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/sort/impl/LocalDateFieldSortBuilderFactory.java
@@ -6,52 +6,20 @@
  */
 package org.hibernate.search.backend.lucene.types.sort.impl;
 
-import java.lang.invoke.MethodHandles;
 import java.time.LocalDate;
 
-import org.hibernate.search.backend.lucene.logging.impl.Log;
 import org.hibernate.search.backend.lucene.search.sort.impl.LuceneSearchSortBuilder;
 import org.hibernate.search.backend.lucene.types.converter.impl.LuceneFieldConverter;
-import org.hibernate.search.engine.logging.spi.EventContexts;
-import org.hibernate.search.engine.search.sort.spi.DistanceSortBuilder;
 import org.hibernate.search.engine.search.sort.spi.FieldSortBuilder;
-import org.hibernate.search.engine.spatial.GeoPoint;
-import org.hibernate.search.util.impl.common.LoggerFactory;
 
-public class LocalDateFieldSortBuilderFactory implements LuceneFieldSortBuilderFactory {
-
-	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
-
-	private final LuceneFieldConverter<LocalDate, ?> converter;
+public class LocalDateFieldSortBuilderFactory extends AbstractStandardLuceneFieldSortBuilderFactory<LocalDate> {
 
 	public LocalDateFieldSortBuilderFactory(LuceneFieldConverter<LocalDate, ?> converter) {
-		this.converter = converter;
+		super( converter );
 	}
 
 	@Override
 	public FieldSortBuilder<LuceneSearchSortBuilder> createFieldSortBuilder(String absoluteFieldPath) {
 		return new LocalDateFieldSortBuilder( absoluteFieldPath, converter );
-	}
-
-	@Override
-	public DistanceSortBuilder<LuceneSearchSortBuilder> createDistanceSortBuilder(String absoluteFieldPath,
-			GeoPoint center) {
-		throw log.distanceOperationsNotSupportedByFieldType(
-				EventContexts.fromIndexFieldAbsolutePath( absoluteFieldPath )
-		);
-	}
-
-	@Override
-	public boolean isDslCompatibleWith(LuceneFieldSortBuilderFactory obj) {
-		if ( this == obj ) {
-			return true;
-		}
-		if ( obj.getClass() != LocalDateFieldSortBuilderFactory.class ) {
-			return false;
-		}
-
-		LocalDateFieldSortBuilderFactory other = (LocalDateFieldSortBuilderFactory) obj;
-
-		return converter.isDslCompatibleWith( other.converter );
 	}
 }
